pyproject.toml
[project]
name = "backtesting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.127.0",
    "gunicorn>=23.0.0",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]

[project.scripts]
backtesting = "backtesting.app:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
